pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aegismesh"
version = "1.4.0"
description = "Sovereign Agent-to-Agent (A2A) Diagnostic Mesh"
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    # ── Web / RPC ──────────────────────────────────────────────────────────
    "fastapi>=0.111.0,<0.116.0",
    "uvicorn[standard]>=0.29.0,<0.32.0",
    "httpx>=0.27.0,<0.29.0",
    "pydantic>=2.7.0,<3.0.0",

    # ── Async helpers ──────────────────────────────────────────────────────
    "aiofiles>=23.2.1,<25.0.0",

    # ── LangGraph / AI Orchestration ──────────────────────────────────────
    # Let pip resolve compatible langchain-core automatically
    "langgraph>=0.2.0,<0.3.0",
    "langchain-core>=1.0.0,<2.0.0",
    "langchain-openai>=1.0.0,<2.0.0",
    "langchain-ollama>=1.0.0,<2.0.0",

    # ── Graph / Evidence Engine ────────────────────────────────────────────
    "networkx>=3.3,<4.0",

    # ── System Monitoring ─────────────────────────────────────────────────
    "psutil>=5.9.8,<7.0.0",

    # ── Observability ─────────────────────────────────────────────────────
    "opentelemetry-api>=1.24.0,<2.0.0",
    "opentelemetry-sdk>=1.24.0,<2.0.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24.0,<2.0.0",

    # ── Utilities ─────────────────────────────────────────────────────────
    "python-dotenv>=1.0.1,<2.0.0",
    "rich>=13.7.1,<15.0.0",
    "typer>=0.12.3,<0.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.2.0",
    "pytest-asyncio==0.23.6",
    "pytest-cov==5.0.0",
    "httpx",                             # already in main, needed for TestClient
    "ruff==0.4.4",
    "mypy==1.10.0",
]

[project.scripts]
aegismesh = "aegismesh.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
